refactor: finish rename by fixing webpack resolution
